Update: Make the backend controllers more maintainable. #531
